fix integer literal precision: parse as uint64 instead of float64
The boson lexer was parsing all numeric literals (including integers)
through strconv.ParseFloat into float64, which silently lost precision
for values above 2^53. Notably, UINT64_MAX = 18446744073709551615 was
rounding to 2^64 and then wrapping to INT64_MIN on conversion to uint64.

Switch token.nval from float64 to uint64 and parse the integer portion
via strconv.ParseUint. Decimal syntax (e.g. 1234.5) is still accepted
by the lexer for grammatical compatibility — the integer portion is
extracted losslessly and the fractional part is silently truncated,
matching prior behavior. When boson grows float support, this is where
the proper handling will go.

The cast_test now uses UINT64_MAX for the u64->i64 reinterpretation
case, which is the value that originally surfaced this bug.

+++ cmd/bosc/lexer_test.go
@@ -20,7 +20,7 @@ func TestLexer(t *testing.T) {
 			out: []token{
 				token{t: tok_ident, sval: "b"},
 				token{t: tok_eq},
-				token{t: tok_byte, nval: float64('h')},
+				token{t: tok_byte, nval: uint64('h')},
 				token{t: tok_ident, sval: "putc"},
 				token{t: tok_lparen},
 				token{t: tok_ident, sval: "b"},
